Refactor ommx tracing modules
