+++ python/ommx/ommx/tracing/_capture.py
@@ -1,9 +1,9 @@
-"""Script-side tracing API: ``capture_trace`` and ``@traced``.
+"""Script-side trace capture context manager.
 
 The Jupyter cell magic (``%%ommx_trace``) is great inside a notebook
 but useless from a plain Python script. This module exposes the same
-``_collector`` / ``_render`` machinery through a context manager and a
-decorator so callers outside IPython get the same tree + JSON output.
+collector through a context manager so callers outside IPython can get
+a :class:`TraceResult`.
 
 Usage::
 
@@ -15,16 +15,6 @@ Usage::
 
     print(trace.text_tree())
     trace.save_chrome_trace("out.json")
-
-Or as a decorator::
-
-    from ommx.tracing import traced
-
-    @traced(output="process.json")
-    def process():
-        ...
-
-    process()  # writes process.json on return *and* on exception
 
 **Exception handling.** Everything inside the managed block still
 raises normally — we never swallow. Before the exception propagates:
@@ -42,28 +32,16 @@ raises normally — we never swallow. Before the exception propagates:
 
 from __future__ import annotations
 
-import functools
-import inspect
 from contextlib import AbstractContextManager
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import Any, Callable, Optional, Union, overload
+from typing import Optional
 
 from opentelemetry import context as otel_context
 from opentelemetry import trace
-from opentelemetry.proto.collector.trace.v1.trace_service_pb2 import (
-    ExportTraceServiceRequest,
-)
-from opentelemetry.proto.trace.v1.trace_pb2 import Span as ProtoSpan
 from opentelemetry.trace import Span
 
-from ._collector import _CellSpanCollector
-from ._otlp import (
-    request_from_otlp_protobuf,
-    request_to_otlp_protobuf,
-    spans_to_otlp_request,
-)
-from ._render import chrome_trace_json, render_text_tree
+from ._collector import _TraceSpanCollector
+from ._otlp import spans_to_otlp_request
+from ._result import TraceResult
 from ._setup import ensure_collector_installed
 
 
@@ -74,58 +52,6 @@ from ._setup import ensure_collector_installed
 # we don't need a per-entry-point tracer.
 _TRACER_NAME = "ommx.tracing"
 _DEFAULT_ROOT_SPAN_NAME = "ommx_trace_block"
-
-
-@dataclass
-class TraceResult:
-    """Populated result of a ``capture_trace`` block.
-
-    Filled in by :class:`capture_trace` on ``__exit__`` (including the
-    exception path, so the caller can always inspect the trace even
-    when the block raised).
-    """
-
-    request: ExportTraceServiceRequest = field(
-        default_factory=ExportTraceServiceRequest
-    )
-
-    @property
-    def spans(self) -> list[ProtoSpan]:
-        """Flattened OTLP protobuf spans exported in this trace result."""
-        return [
-            span
-            for resource_span in self.request.resource_spans
-            for scope_span in resource_span.scope_spans
-            for span in scope_span.spans
-        ]
-
-    @classmethod
-    def from_otlp_protobuf(cls, payload: bytes) -> "TraceResult":
-        """Build a trace result from an OMMX trace layer payload."""
-        return cls(request=request_from_otlp_protobuf(payload))
-
-    def text_tree(self) -> str:
-        """Return the nested text tree — same renderer the cell magic uses."""
-        return render_text_tree(self.request)
-
-    def otlp_protobuf(self) -> bytes:
-        """Return OTLP protobuf bytes stored in Experiment trace layers."""
-        return request_to_otlp_protobuf(self.request)
-
-    def chrome_trace_json(self) -> str:
-        """Return a Chrome Trace Event Format JSON string."""
-        return chrome_trace_json(self.request)
-
-    def save_chrome_trace(self, path: Union[str, Path]) -> None:
-        """Write the Chrome Trace JSON to ``path`` (creating parents as needed).
-
-        Overwrites any existing file. The UTF-8 encoding matches the
-        JSON spec and is what Perfetto / speedscope /
-        ``chrome://tracing`` all accept.
-        """
-        p = Path(path)
-        p.parent.mkdir(parents=True, exist_ok=True)
-        p.write_text(self.chrome_trace_json(), encoding="utf-8")
 
 
 class capture_trace:  # noqa: N801 - context-manager factory, lowercase on purpose
@@ -142,7 +68,7 @@ class capture_trace:  # noqa: N801 - context-manager factory, lowercase on purpo
         self._name = name
         self._result = TraceResult()
         self._trace_id: Optional[int] = None
-        self._collector: Optional[_CellSpanCollector] = None
+        self._collector: Optional[_TraceSpanCollector] = None
         self._span_cm: Optional[AbstractContextManager[Span]] = None
 
     def __enter__(self) -> TraceResult:
@@ -177,135 +103,3 @@ class capture_trace:  # noqa: N801 - context-manager factory, lowercase on purpo
                 self._result.request = spans_to_otlp_request(spans)
         # Never swallow the exception.
         return False
-
-
-# ---------------------------------------------------------------------------
-# @traced decorator
-# ---------------------------------------------------------------------------
-
-
-_F = Callable[..., Any]
-
-
-@overload
-def traced(func: _F) -> _F: ...
-
-
-@overload
-def traced(
-    *,
-    name: Optional[str] = ...,
-    output: Optional[Union[str, Path]] = ...,
-) -> Callable[[_F], _F]: ...
-
-
-def traced(
-    func: Optional[_F] = None,
-    *,
-    name: Optional[str] = None,
-    output: Optional[Union[str, Path]] = None,
-) -> Any:
-    """Decorator that runs the wrapped function under :class:`capture_trace`.
-
-    Supports all three call shapes::
-
-        @traced
-        def process(): ...
-
-        @traced()
-        def process(): ...
-
-        @traced(name="build_qubo", output="qubo.json")
-        def process(): ...
-
-    If ``output`` is given, the Chrome Trace JSON is written to that
-    path when the function returns **or raises** — information is
-    never dropped. The exception, if any, is re-raised unchanged after
-    the file is written.
-
-    If ``name`` is omitted, the span is named after the function
-    (``fn.__qualname__``) so traces from multiple decorated functions
-    are easy to tell apart in the rendered tree.
-    """
-
-    def _save_if_configured(result: Optional[TraceResult]) -> None:
-        if output is not None and result is not None:
-            result.save_chrome_trace(output)
-
-    def _save_best_effort(result: Optional[TraceResult]) -> None:
-        """Like ``_save_if_configured`` but swallows any I/O failure.
-
-        Used on the exception path: a save failure here would *replace*
-        the user's original exception, which is the signal they care
-        about most. Silently dropping the save is the lesser evil.
-        """
-        if output is None or result is None:
-            return
-        try:
-            result.save_chrome_trace(output)
-        except Exception:  # noqa: BLE001 - intentional swallow
-            pass
-
-    def _decorator(fn: _F) -> _F:
-        span_name = name if name is not None else fn.__qualname__
-
-        if inspect.iscoroutinefunction(fn):
-            # ``async def`` needs its own wrapper: a plain sync wrapper
-            # would trace only the coroutine-object creation, finish
-            # the ``capture_trace`` block, and return the still-
-            # unawaited coroutine — by the time it runs, the capture
-            # window is closed and every span is silently dropped.
-            @functools.wraps(fn)
-            async def _async_wrapper(*args, **kwargs):
-                capture = capture_trace(span_name)
-                result: Optional[TraceResult] = None
-                # Pre-initialise so pyright can see ``retval`` is
-                # bound in the ``else`` branch even though the
-                # assignment below lives inside a ``with`` whose
-                # ``__exit__`` could theoretically raise.
-                retval: Any = None
-                try:
-                    with capture as r:
-                        result = r
-                        retval = await fn(*args, **kwargs)
-                except BaseException:
-                    # User exception: save best-effort, then propagate
-                    # it unchanged so the signal isn't lost to an I/O
-                    # hiccup from the trace writer.
-                    _save_best_effort(result)
-                    raise
-                else:
-                    # Success: let save errors surface so a broken
-                    # path configuration is noticed.
-                    _save_if_configured(result)
-                    return retval
-
-            return _async_wrapper  # type: ignore[return-value]
-
-        @functools.wraps(fn)
-        def _wrapper(*args, **kwargs):
-            capture = capture_trace(span_name)
-            result: Optional[TraceResult] = None
-            retval: Any = None
-            try:
-                with capture as r:
-                    result = r
-                    retval = fn(*args, **kwargs)
-            except BaseException:
-                # User exception: save best-effort, then propagate
-                # it unchanged so the signal isn't lost to an I/O
-                # hiccup from the trace writer.
-                _save_best_effort(result)
-                raise
-            else:
-                # Success: let save errors surface so a broken path
-                # configuration is noticed.
-                _save_if_configured(result)
-                return retval
-
-        return _wrapper
-
-    if func is not None:
-        # ``@traced`` without parens.
-        return _decorator(func)
-    return _decorator
