+++ python/ommx/ommx/tracing/_magic.py
@@ -23,10 +23,16 @@ HTML rendering with the ``[ERROR]`` marker) comes from
 
 from __future__ import annotations
 
+import base64
+import html
 from typing import TYPE_CHECKING, Optional, Tuple
 
+from opentelemetry.proto.collector.trace.v1.trace_service_pb2 import (
+    ExportTraceServiceRequest,
+)
+
 from ._capture import capture_trace
-from ._render import render_cell_output_html
+from ._render import chrome_trace_json, render_text_tree
 
 
 if TYPE_CHECKING:  # pragma: no cover - type hints only
@@ -34,6 +40,29 @@ if TYPE_CHECKING:  # pragma: no cover - type hints only
 
 
 _CELL_ROOT_SPAN_NAME = "ommx_trace_cell"
+
+
+def _render_cell_output_html(
+    request: ExportTraceServiceRequest,
+    *,
+    download_filename: str = "ommx_trace.json",
+) -> str:
+    """HTML blob for ``display(HTML(...))`` from the cell magic."""
+    tree = html.escape(render_text_tree(request))
+    payload = chrome_trace_json(request)
+    b64 = base64.b64encode(payload.encode("utf-8")).decode("ascii")
+    data_url = f"data:application/json;base64,{b64}"
+    size_kb = len(payload) / 1024
+    # ``quote=True`` escapes both ``"`` and ``'`` for HTML attributes.
+    safe_filename = html.escape(download_filename, quote=True)
+    return (
+        '<div class="ommx-trace">'
+        f"<pre>{tree}</pre>"
+        f'<p><a href="{data_url}" download="{safe_filename}">'
+        f"Download Chrome Trace JSON ({size_kb:.1f} KB)"
+        "</a> — open in Perfetto, speedscope, or <code>chrome://tracing</code>.</p>"
+        "</div>"
+    )
 
 
 def run_cell_with_trace(
@@ -93,7 +122,7 @@ def run_cell_with_trace(
             # spec's intent for this flag.
             root.record_exception(cell_exc, escaped=True)
 
-    return render_cell_output_html(trace_result.request), cell_exc
+    return _render_cell_output_html(trace_result.request), cell_exc
 
 
 def register_magic(shell: "InteractiveShell") -> None:
